Resolves conflicts with 35 upstream commits, principally the fused-MoE
CuTe DSL rework that adds MXFP8 x MXFP4 support.

The substantive upstream change is that the Blackwell kernels' can_implement
signatures split `ab_dtype` into separate `a_dtype`/`b_dtype` (A and B may now
differ). Our SM107 kernels already used the split form, so the resolution keeps
our Rubin/Blackwell dispatch and adopts the new signature on the Blackwell side:

- tuner.py: kept the `_is_rubin_tactic` branch; updated both Blackwell
  can_implement calls to a_dtype/b_dtype.
- act_fusion / finalize_fusion: kept the `is_rubin` branches and the
  arch-specific tactic parameters (3D `mma_tiler` + `mma_inst_shape` for SM107
  alongside Blackwell's `mma_tiler_mn`), unioned them with upstream's new dtype
  parameters, and renamed our `ab_dtype_cutlass` uses to `a_dtype_cutlass` /
  `b_dtype_cutlass` to match the auto-merged regions.
- Cache keys keep the "sm107"/"sm100" discriminator and gain upstream's split
  dtypes.
- __init__ docstring: unioned (upstream's dtype list, our Rubin mention).

Also threads `mma_tiler`/`mma_inst_shape` through the `*_nvfp4` convenience
wrappers. Upstream restructured those into back-compat shims that accept only
`mma_tiler_mn`, so the Rubin tactic parameters had no path from fused_moe.py to
the kernels; mypy caught this as an unexpected-keyword error rather than it
surfacing at runtime.

+++ flashinfer/fused_moe/cute_dsl/tuner.py
@@ -733,7 +733,8 @@ class CuteDslFusedMoENvfp4Runner(TunableRunner):
                 gemm2_mma_tiler_mn, gemm2_cluster_shape_mn, _ = gemm2_tactic
 
                 gemm1_ok = BlockScaledContiguousGatherGroupedGemmKernel.can_implement(
-                    ab_dtype=ab_dtype,
+                    a_dtype=ab_dtype,
+                    b_dtype=ab_dtype,
                     sf_dtype=sf_dtype,
                     sf_vec_size=sf_vec_size,
                     c_dtype=gemm1_c_dtype,
@@ -748,7 +749,8 @@ class CuteDslFusedMoENvfp4Runner(TunableRunner):
                     c_major="n",
                 )
                 gemm2_ok = Sm100BlockScaledContiguousGroupedGemmFinalizeFusionKernel.can_implement(
-                    ab_dtype=ab_dtype,
+                    a_dtype=ab_dtype,
+                    b_dtype=ab_dtype,
                     sf_dtype=sf_dtype,
                     sf_vec_size=sf_vec_size,
                     out_dtype=gemm2_out_dtype,
